# openapi.yml
...
openapi: 3.1.0
info:
  title: Postmill API (reaatech fork)
  version: 3.4.0
  description: |
    Comprehensive OpenAPI description of the **Postmill** backend (NestJS) as it exists in this
    heavily-modified `reaatech` fork.

    Postmill schedules social-media and chat posts to 28+ channels, with a calendar, analytics, team
    management, a media library, and — in this fork — a pluggable multi-provider **AI layer**,
    **Analytics v2** persisted dashboards, **social comments** sync/moderation, admin-configurable
    **channel OAuth credentials**, and an OAuth2 authorization-code provider.

    ## Surfaces & authentication

    There are three distinct authentication surfaces:

    1. **Dashboard API** (the bulk of routes). Requires a Postmill user session JWT supplied either as
       the `auth` **cookie** or the `auth` **request header**. The active organization is selected
       with the `showorg` cookie/header. Many routes are further gated by RBAC policies
       (`@CheckPolicies`) or by `isSuperAdmin`.
    2. **Public integration API** (`/public/v1/*`). Requires an organization **API key** (or a
       `pos_`-prefixed OAuth access token) in the `Authorization` header. Intended for
       n8n/Zapier/CLI/3rd-party automation.
    3. **Unauthenticated routes** — registration/login, the OAuth2 authorization/token endpoints,
       Stripe & enterprise webhooks, the social-connect OAuth callback, health/monitor, and public
...

...

    ## Rate limiting
...

  contact:
...
  license:
...
servers:
...
security:
  - cookieAuth: []
  - authHeader: []
paths:
...
  /posts:
    get:
      tags: [Posts]
      summary: Calendar posts within a date range
      operationId: getPosts
      parameters:
        - { name: startDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: endDate, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: customer, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Minified posts for the calendar
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/CalendarPost' } }
    post:
      tags: [Posts]
      summary: Create / schedule / draft a post group
      operationId: createPost
      description: |
        Server-side validates every channel's content & settings. Non-draft posts that fail
        validation return a structured `400` (see `PostValidationError`).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePostDto' }
      responses:
        '200':
          description: Created post group
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CreatePostResult' }
        '400': { $ref: '#/components/responses/PostValidation' }
        '406': { $ref: '#/components/responses/SubscriptionRequired' }
...
  /posts/inbox:
    get:
      tags: [Social Comments]
      summary: Cross-channel comment inbox
      operationId: getCommentInbox
      parameters:
        - { name: status, in: query, schema: { type: string } }
        - { name: assigneeId, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: unreadOnly, in: query, schema: { type: string, enum: ['true', 'false'] } }
      responses: { '200': { description: Inbox page } }
...
  /posts/inbox/bulk-read:
    post:
      tags: [Social Comments]
      summary: Bulk mark comments as read
      operationId: inboxBulkRead
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [commentIds], properties: { commentIds: { type: array, items: { type: string } } } }
      responses: { '200': { description: Marked } }
  /posts/{id}/social-comments:
    get:
      tags: [Social Comments]
      summary: List synced platform comments for a post
      operationId: getSocialComments
      parameters:
        - { $ref: '#/components/parameters/PathUuid' }
        - { name: cursor, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Paginated comments with per-user read state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SocialCommentsPage' }
    post:
      tags: [Social Comments]
      summary: Reply to the post itself
      operationId: addSocialComment
      parameters: [{ $ref: '#/components/parameters/PathUuid' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplyCommentDto' } } }
      responses: { '200': { description: Reply result, content: { application/json: { schema: { type: object, additionalProperties: true } } } } }
  /posts/{id}/social-comments/read:
    post:
      tags: [Social Comments]
      summary: Mark all of a post's comments as read
      operationId: markCommentsRead
      parameters: [{ $ref: '#/components/parameters/PathUuid' }]
      responses: { '200': { description: Marked read } }
  /posts/{id}/social-comments/unread-count:
    get:
      tags: [Social Comments]
      summary: Unread comment count for a post
      operationId: getUnreadCommentCount
      parameters: [{ $ref: '#/components/parameters/PathUuid' }]
      responses:
        '200': { description: Count, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
  /posts/{id}/social-comments/{commentId}/reply:
    post:
      tags: [Social Comments]
      summary: Reply to a specific comment
      operationId: replyToComment
      parameters:
        - { $ref: '#/components/parameters/PathUuid' }
        - { $ref: '#/components/parameters/PathCommentId' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ReplyCommentDto' } } }
      responses: { '200': { description: Reply result } }
  /posts/{id}/social-comments/{commentId}/like:
    post:
      tags: [Social Comments]
      summary: Like / unlike a comment
      operationId: likeComment
      parameters:
        - { $ref: '#/components/parameters/PathUuid' }
        - { $ref: '#/components/parameters/PathCommentId' }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LikeCommentDto' } } }
      responses: { '200': { description: Like result } }
  /posts/{id}/social-comments/{commentId}/status:
    post:
      tags: [Social Comments]
      summary: Update a comment's moderation status
      operationId: updateCommentStatus
      parameters:
        - { $ref: '#/components/parameters/PathUuid' }
        - { $ref: '#/components/parameters/PathCommentId' }
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { status: { type: string } } } } }
      responses: { '200': { description: Updated } }
  /posts/{id}/social-comments/{commentId}/assign:
    post:
      tags: [Social Comments]
      summary: Assign a comment to a team member
      operationId: assignComment
      parameters:
        - { $ref: '#/components/parameters/PathUuid' }
        - { $ref: '#/components/parameters/PathCommentId' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { assigneeId: { type: [string, 'null'], description: User id, or null to unassign } }
      responses: { '200': { description: Assigned } }

  # ===================================================================== CAMPAIGNS (fork)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: Session JWT in the `auth` cookie (dashboard API).
    authHeader:
      type: apiKey
      in: header
      name: auth
      description: Session JWT in the `auth` header (alternative to the cookie).
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        Organization **API key** or a `pos_`-prefixed OAuth access token, sent **raw** (no `Bearer`
        prefix) in the `Authorization` header. Used by `/public/v1/*`.
  parameters:
...
    PathUuid:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PathCommentId:
      name: commentId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    SubscriptionRequired:
      description: An active subscription is required for this action
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    PostValidation:
      description: Per-channel post validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/PostValidationError' }
  schemas:
...
    Error:
      type: object
      properties:
        statusCode: { type: integer, example: 400 }
        message:
          oneOf:
            - { type: string }
            - { type: array, items: { type: string } }
          example: Bad Request
        error: { type: string, example: Bad Request }
...
    PostValidationError:
      type: object
      description: Structured per-channel validation failure (PostValidationExceptionFilter).
      properties:
        provider: { type: string, example: twitter }
        name: { type: string, example: My Twitter }
        error: { type: string, example: post is too long, please fix it }

    # ---------- Auth DTOs ----------
...
    MediaRef:
      type: object
      properties:
        id: { type: string }
        path: { type: string }
      additionalProperties: true
    PostContent:
      type: object
      required: [content, image]
      properties:
        content: { type: string, description: Sanitized post content }
        id: { type: string }
        delay: { type: integer, description: Delay (ms) for thread items }
        image:
          type: array
          items: { $ref: '#/components/schemas/MediaRef' }
    Post:
      type: object
      required: [integration, value]
      properties:
        type: { type: string }
        integration:
          type: object
          required: [id]
          properties: { id: { type: string } }
        value:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/PostContent' }
        group: { type: string }
        settings:
          type: object
          description: Provider-specific settings (discriminated by `__type`).
          additionalProperties: true
    CreatePostDto:
      type: object
      required: [type, shortLink, date, tags, posts]
      properties:
        type: { type: string, enum: [draft, schedule, now, update] }
        order: { type: string }
        shortLink: { type: boolean }
        inter: { type: number }
        date: { type: string, format: date-time }
        tags:
          type: array
          items:
            type: object
            properties: { value: { type: string }, label: { type: string } }
        posts:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Post' }
      example:
        type: schedule
        shortLink: false
        date: '2026-06-10T14:00:00.000Z'
        tags: []
        posts:
          - integration: { id: clt_integration_123 }
            value:
              - content: 'Hello from the Postmill API 👋'
                image: []
            settings: { __type: twitter }
    CreatePostResult:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        group: { type: string }
...
    CalendarPost:
      type: object
      additionalProperties: true
      properties:
        id: { type: string }
        content: { type: string }
        publishDate: { type: string, format: date-time }
        state: { type: string, example: QUEUE }
        integration: { type: object, additionalProperties: true }
...
    ReplyCommentDto:
      type: object
      required: [message]
      properties:
        message: { type: string, maxLength: 10000 }
    LikeCommentDto:
      type: object
      required: [like]
      properties:
        like: { type: boolean }
    SocialComment:
      type: object
      additionalProperties: true
      properties:
        id: { type: string, format: uuid }
        content: { type: string }
        authorName: { type: string }
        status: { type: string, example: open }
        assigneeId: { type: [string, 'null'] }
        read: { type: boolean }
        createdAt: { type: string, format: date-time }
    SocialCommentsPage:
      type: object
      properties:
        comments:
          type: array
          items: { $ref: '#/components/schemas/SocialComment' }
        nextCursor: { type: [string, 'null'] }
      additionalProperties: true

    # ---------- Integrations ----------
...